Iteration on Skeletal Mesh Particles
Now using "BP_DeltaTime" to control the flow of animation and to switch targets - is working relatively well.

Next step - consider using 4 emitters: one for the inActive positions, one for the Active positions, and another for the beam, and then a final one for the rendered particles that blends between each of the others based on state.